+++ scripts/test/Selenium/Output/Dashboard/TicketGenericFilter.t
@@ -44,6 +44,11 @@ $Selenium->RunTest(
                 CustomerUser => $Helper->GetRandomID() . '@third.com',
                 CustomerID   => 'CustomerCompany' . $Helper->GetRandomID() . '(#)',
             },
+            {
+                TN           => $TicketObject->TicketCreateNumber(),
+                CustomerUser => $Helper->GetRandomID() . '@fourth.com',
+                CustomerID   => 'CustomerCompany#%' . $Helper->GetRandomID() . '(#)',
+            },
         );
 
         # Create test tickets.
@@ -341,6 +346,37 @@ $Selenium->RunTest(
                     'return typeof($) === "function" && !$("#Dashboard0120-TicketNew-box.Loading").length'
             );
 
+            # Verify if CustomerID containing special characters is filtered correctly. See bug#14982.
+            $Selenium->find_element("//a[contains(\@title, \'Customer ID\' )]")->click();
+            $Selenium->WaitFor(
+                JavaScript =>
+                    "return typeof(\$) === 'function' && \$('.CustomerIDAutoComplete:visible').length;"
+            );
+
+            $Selenium->find_element( ".CustomerIDAutoComplete", 'css' )->send_keys( $Tickets[3]->{CustomerID} );
+
+            # Wait for AJAX to finish.
+            $Selenium->WaitFor(
+                JavaScript =>
+                    'return typeof($) === "function" && !$("span.AJAXLoader:visible").length'
+            );
+
+            # Choose the value.
+            $Selenium->execute_script(
+                "\$('#ColumnFilterCustomerID0120-TicketNew').val('$Tickets[3]->{CustomerID}').trigger('change');"
+            );
+
+            # Wait for autocomplete action.
+            $Selenium->WaitFor(
+                JavaScript =>
+                    "return typeof(\$) === 'function' && \$('td div[title=\"$Tickets[3]->{CustomerID}\"]').length == 1;"
+            );
+
+            # Verify CustomerID containing special characters is found.
+            $Self->True(
+                index( $Selenium->get_page_source(), $Tickets[3]->{TN} ) > -1,
+                "Test ticket with TN $Tickets[3]->{TN} - found on screen after filtering with customer - $Tickets[3]->{CustomerID}",
+            );
         }
 
         # Delete test tickets.